Edit div sections and add styling to header
I removed an unnecessary div tags in the HTML file and created a
new class for the unordered list elements. I also used flexbox to
display the header links aligned horizontally on the header. I used
margins and gaps to create an appropriate amount of spacing for the
header logo and the header links.