Edit second subsection classes and spacing
I changed some of the classes used to display the image/illustration
placeholders and the subtext. Some of the class changes made it easier
to properly align the elements with one another. I used flexbox to
align the image/illustration placeholders and the subtext. Flexbox was
also used to horizontally center and properly space the elements in the
second subsection.